In the Scheduler interface, split runNextTask(bool sleep) into runNextTask() and runNextTaskOrSleep().  Not parameterizing the function call alleviates the need for GCC to correctly eliminate dead code, minimizing code bloat.
Also, slightly changed the semantics of runNextTaskOrSleep() to not spin until
a task is run.  This wasn't necessary and the function returns a bool to
indicate a task didn't run if it is ever pertinent (which it currently never
is).

In SchedulerM, changed all explicit management of atomic state using the
__nesc functions to "plain" atomic blocks which are more clear.  This change
in postTask was also the source of the biggest reduction of compiled code
bloat.  Also removed atomic blocks from internal functions, leaving that as
the responsibility of the interface commands themselves.

For a little more clarity, tweaked the TaskBasic interface to postTask() and
runTask() from post_() and run().

These changes all together reduced apps/Blink from 3556 ROM bytes to 3138 ROM
bytes for telosb, a 418 ROM byte savings.