Significant reduction in the number of interface.  All the *Base interfaces have been removed in favor of defaults.  In part, this means
  - Only one Timer interface instead of four.  It is synchronous, 32-bits, and
    parameterized by frequency.

  - Only one Alarm interface instead of two.  It is asynchronous and
    parameteriezd by frequency and size.

This also means none of the Cast* components are necessary and have been
removed.

All single modules have been renamed from *M to *C, where "C" is taken to mean
component not configuration.  This is the best naming convention as proposed
by Ion, leaving room for a *C/*M pair to exist if ever need be without
changing components that depend on them.

Added BusyWait and LocalTime interfaces.  LocalTime is a 32-bit Counter with
the overflow event hidden.

Renamed MultiplexTimerM to VirtualizeTimerC.