Revert "DMA-driven UartStream for msp430."
This reverts commit 59fbd1b82c12cd0ed07b0886d46215deaead8399.

The approach taken in that commit supports DMA for RX only, not TX;
hard-codes a dependency on DMA channel 2, and changes the behavior of the
UartStream buffer-based receive infrastructure to produce a stream of
receiveDone events rather than the standard one event per receive command.
An alternative approach will be used to solve the problems this patch was
intended to address.